fix(studio): keep the canvas iframe out of the editor shell
`check-studio-dist` caught this, which is what it is for: importing
`resolveAssetRef` from `asset-refs.ts` dragged the entire editor shell into the
iframe bundle, because that module reaches `projectState`, `activeTab` and the
registered platform. The visible symptom was a stray `dist/iframe-entry.css`
nothing accounted for; the real one is that `iframe-render.ts` says in its own
header that it stays dependency-light, and it had stopped.

So the split is by REALM rather than by taste. `asset-resolve.ts` is a pure
function of its arguments and is what the iframe imports; `asset-refs.ts` is
everything that has to ask the running editor a question, and re-exports the
pure half so a parent-realm caller still imports one module.

The dep-rules gate was right too, and for a funnier reason: its regex reads
`from` followed by a quoted package name as an import, and a comment saying
"Duplicated from `@jxsuite/parser`'s content loader" is exactly that shape. The
comment now names the file rather than the package — and says the other half of
why the constant is duplicated, which is that core may not import an extension
at all.
